Renamed to universal3D_preparation.
