+++ tests/test_3D_unit.py
@@ -1,4 +1,4 @@
-from autogaita.autogaita_3D_preparation import clean_a_file, rename_a_column
+from autogaita.autogaita_universal3D_preparation import clean_a_file, rename_a_column
 from hypothesis import given, strategies as st
 import pandas as pd
